+++ examples/storybook/stories/Plugins/index.js
@@ -1,4 +1,5 @@
 import { storiesOf } from '@storybook/react';
+import LinkPreview from './LinkPreview';
 import GalleryPlugin from './GalleryPlugin.js';
 import DividerPluginStory from './DividerPluginStory.js';