+++ examples/main/shared/editor/Editor.jsx
@@ -7,7 +7,6 @@ import { testImages, testVideos } from './mock';
 import * as Plugins from './EditorPlugins';
 import ModalsMap from './ModalsMap';
 import theme from '../theme/theme'; // must import after custom styles
-import { debugBiLoggers } from '../../config/biService';
 import { GALLERY_TYPE } from 'wix-rich-content-plugin-gallery';
 
 const modalStyleDefaults = {
@@ -55,7 +54,12 @@ export default class Editor extends PureComponent {
       }
     };
     this.helpers = {
-      ...debugBiLoggers,
+      //these are for testing purposes only
+      onPluginAdd: async (plugin_id, entry_point, version) => console.log('biPluginAdd', plugin_id, entry_point, version),
+      onPluginDelete: async (plugin_id, version) => console.log('biPluginDelete', plugin_id, version),
+      onPluginChange: async (plugin_id, changeObj, version) => console.log('biPluginChange', plugin_id, changeObj, version),
+      onPublish: async (postid, callback, version) => console.log('biOnPublish', ({ data }) => data, version),
+      //
       // onFilesChange: (files, updateEntity) => mockUpload(files, updateEntity),
       handleFileSelection: (index, multiple, updateEntity, removeEntity, componentData) => {
         const count = componentData.items || shouldMultiSelectImages ? [1, 2, 3] : [1];