+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -228,7 +228,7 @@ const videoHandlers = {
       updateEntity({ data: videoToUpload });
       //updateEntity({ error: { msg: 'Upload Failed' } });
       console.log('consumer uploaded ', videoToUpload);
-    }, 1500);
+    }, 1200000);
   },
 };