+++ examples/main/shared/viewer/Viewer.jsx
@@ -5,8 +5,10 @@ import { isSSR } from 'wix-rich-content-common';
 import * as PropTypes from 'prop-types';
 import * as Plugins from './ViewerPlugins';
 import theme from '../theme/theme'; // must import after custom styles
-import getImagesData from 'wix-rich-content-fullscreen/src/lib/getImagesData';
+import getImagesData from 'wix-rich-content-fullscreen/dist/lib/getImagesData';
 import Fullscreen from 'wix-rich-content-fullscreen';
+import 'wix-rich-content-fullscreen/dist/styles.min.css';
+
 import {
   TextSelectionListener,
   ViewerInlineToolBar,
@@ -42,20 +44,17 @@ export default class Viewer extends PureComponent {
     onExpand: (entityIndex, innerIndex = 0) => {
       //galleries have an innerIndex (i.e. second image will have innerIndex=1)
       this.setState({
-        expendModeIsOpen: true,
+        expandModeIsOpen: true,
         expandModeIndex: this.expandModeData.imageMap[entityIndex] + innerIndex,
       });
     },
   };
 
   render() {
-    const { isMobile, locale, initialState, seoMode } = this.props;
-    const { expendModeIsOpen, expandModeIndex, disabled } = this.state;
+    const { isMobile, initialState, seoMode } = this.props;
+    const { expandModeIsOpen, expandModeIndex, disabled } = this.state;
 
     const viewerProps = {
-      locale,
-      relValue,
-      anchorTarget,
       isMobile,
       theme,
       initialState,
@@ -77,9 +76,9 @@ export default class Viewer extends PureComponent {
           />
           {!isSSR() && (
             <Fullscreen
-              isOpen={expendModeIsOpen}
               images={this.expandModeData.images}
-              onClose={() => this.setState({ expendModeIsOpen: false })}
+              onClose={() => this.setState({ expandModeIsOpen: false })}
+              isOpen={expandModeIsOpen}
               index={expandModeIndex}
             />
           )}