+++ CHANGELOG.md
@@ -28,6 +28,9 @@
   - [#757](https://github.com/wix-incubator/rich-content/pull/757) prevent underline link inline style to affects on new line
 - `link`
   - [#756](https://github.com/wix-incubator/rich-content/pull/756) selection keeps in place when canceling link panel
+- `gallery`
+  - [#697](https://github.com/wix-incubator/rich-content/pull/697) altText keeping value when changing to next image 
+
 </details>
 <hr/>