+++ CHANGELOG.md
@@ -14,8 +14,13 @@
   </summary>
 
 ### :rocket: New Feature
+- `gallery`
+  - [#625](https://github.com/wix-incubator/rich-content/pull/625) gallery supports adding videos
+- `fullscreen`
+  - [#625](https://github.com/wix-incubator/rich-content/pull/625) supports target prop to render fullscreen at the target element
 - `editor-common`
   - [#716](https://github.com/wix-incubator/rich-content/pull/716) progress loader
+  
 ### :bug: Bug Fix
 - `hashtag`
   - [#707](https://github.com/wix-incubator/rich-content/pull/707) allow hashtag and link in the same block
@@ -29,7 +34,11 @@
 - `link`
   - [#756](https://github.com/wix-incubator/rich-content/pull/756) selection keeps in place when canceling link panel
 - `gallery`
-  - [#697](https://github.com/wix-incubator/rich-content/pull/697) altText keeping value when changing to next image 
+  - [#697](https://github.com/wix-incubator/rich-content/pull/697) altText keeping value when changing to next image
+
+### :house: Internal
+- `fullscreen`
+  - [#625](https://github.com/wix-incubator/rich-content/pull/625) replaced react-images with Pro-Gallery
 
 </details>
 <hr/>