+++ CHANGELOG.md
@@ -15,10 +15,13 @@
   
 ### :boom: New Feature
 - `text-selection-toolbar`
-  - [#523](https://github.com/wix-incubator/rich-content/pull/523) create viewer toolbar with tweet option  
+  - [#523](https://github.com/wix-incubator/rich-content/pull/523) create viewer toolbar with tweet option
 ### :bug: Bug Fix
 - `html`
   - [#732](https://github.com/wix-incubator/rich-content/pull/732) fix html width in viewer
+### :house: Internal
+- `e2e`
+  - [#737](https://github.com/wix-incubator/rich-content/pull/737) fix rendering and plugins tests
     
 </details>
 <hr/>